refactor(empty): 优化empty组件
