+++ components/empty/style/empty.less
@@ -16,7 +16,6 @@
     --f-empty-text-color: var(--f-text-color-secondary);
     --f-empty-description-margin-top: @empty-description-margin-top;
     --f-empty-bottom-margin-top: @empty-bottom-margin-top;
-    --f-svg-monochrome-grey: @empty-svg-monochrome-grey;
 
     display: flex;
     flex-direction: column;
@@ -40,7 +39,6 @@
         svg {
             width: 100%;
             height: 100%;
-            color: var(--f-svg-monochrome-grey);
             vertical-align: top;
             fill: currentcolor;
         }
