+++ components/locales/lang/zhCN.ts
@@ -6,6 +6,9 @@ export default {
         placeholder: '请选择',
         tagOption: '自定义',
     },
+    empty: {
+        emptyText: '暂无数据',
+    },
     cascader: {
         loadingRequiredMessage: '加载全部 {label} 的子节点后才可选中',
     },
