+++ docs/.vitepress/components/empty/custom.vue
@@ -1,9 +1,9 @@
 <template>
-    <FEmpty :image="url" :image-size="80" description="自定义图片" />
+    <FEmpty :image-src="url" description="自定义图片" />
 </template>
 
 <script lang="ts" setup>
 import { ref } from 'vue';
 
-const url = ref('/images/empty-simple.svg');
+const url = ref('/images/empty.svg');
 </script>
