+++ docs/.vitepress/components/empty/default.vue
@@ -1,3 +1,3 @@
 <template>
-    <FEmpty description="暂无数据" />
+    <FEmpty />
 </template>
