+++ docs/.vitepress/components/empty/select.vue
@@ -1,7 +1,7 @@
 <template>
     <FSelect>
         <template #empty>
-            <FEmpty :image-size="80" description="暂无数据" />
+            <FEmpty :image-style="{ width: '80px' }" description="暂无数据" />
         </template>
     </FSelect>
 </template>
