+++ docs/.vitepress/components/empty/size.vue
@@ -1,3 +1,3 @@
 <template>
-    <FEmpty :image-size="120" description="暂无数据" />
+    <FEmpty :image-style="{ width: '240px' }" description="暂无数据" />
 </template>
